Updated vis-timeline.service.ts
I've tried to update the vis timeline service to give the ability to use any of the events declared in the vis timeline documentation. Right now, through ngx-vis, you can only use the events specified in the VisTimelineService class, this is fine for common use cases but if you need to use other events (e.g. "mouseOver" or "drop") you'll have an hard time figuring out how to do so.
I already know that my implementation of the feature is not perfect because a new "EventEmitter" is created every time you call the "on" method but if you like the idea i could work on this a bit more and fix it by creating all the necessary event emitters beforehand.
The nice part is that the old implementation (using `visTimelineService.click.subscribe(() => { ... })`) will still work 😊.
I hope you'll consider this pull request.
Thanks in advice.